feat : 6강

+++ Assets/Script/Projectile.cs
@@ -4,9 +4,10 @@ public class Projectile : MonoBehaviour
 {
     private Movement2D movement2D;
     private Transform target;
-    private int damage;
+    //private int damage;
+    private float damage;
 
-    public void Setup(Transform _target, int _damage)
+    public void Setup(Transform _target, float _damage)
     {
         movement2D = GetComponent<Movement2D>();
         this.target = _target;
